pyproject.toml
[tool.poetry]
name = "my-rag-quiz-app"
version = "0.1.0"
description = ""
readme = "README.md"
packages = [{include = "my_rag_quiz_app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12" # Ou la version que vous avez définie précédemment
flask = "^3.1.0"
langgraph = "^0.3.20"
langchain = "^0.3.21"
langchain-google-genai = "^2.1.1"
langchain-community = "^0.3.20"
faiss-cpu = "^1.10.0"
pymupdf = "^1.25.4"
python-dotenv = "^1.1.0"
langchain-text-splitters = "^0.3.7"
tavily-python = "^0.5.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
